PresenceServer/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
redis[asyncio]
python-jose[cryptography]
